Implemented indexing of annotations in the new 5.0 live-indexing framework.
The indexes get built, but we're currently getting an AbstractMethodError when closing the mention indexes. This sounds like a dereliction of duty on the part of the compiler, but its something to look into some other time, whilst in possession of a clear head.

Changed all ANT build files to target Java 7.