Create annotations from the "entities" property of Twitter JSON
In Twitter JSON the property "entities" is a map from entity type to list of
entities, where each entity is an object with an "indices" property giving its
offsets and other properties giving features.  I've added support for turning
this structure into annotations in the Original markups set.

The DocumentFormat will have this behaviour enabled by default, the populator
currently not (principle of least surprise - I'll enable it once I've added a
checkbox to the GUI to control it).