Upgraded to Grails 2.5.3, as Grails 2.2.x will not start up when run on Java 8.  I still intend to build Mimir on Java 7 for the moment, but at least it will now be possible to run it on 8.
Note that this changes the build process a bit, as Grails 2.5 dependency resolver didn't want to play nicely with dumping JAR files in mimir-web/lib, so I've made Ivy create POM files for mimir-core and mimir-client and install the built JAR files into ~/.m2/repository.  This tidies up the Grails-side BuildConfig considerably as we don't need to duplicate the Ivy dependencies in there - it picks them all up automatically as transitive deps of mimir-core.